Re-organize hud code to separate file / small formatting fixes
- Renamed DrawNearestBase to DrawRangefinder
- Changed formatting in main for terrain and lander files
